requirements.txt
# Memory Chat API Dependencies

# FastAPI and Server
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.0.0

# Zep Memory
zep-cloud>=1.0.0

# Voyage AI for embeddings and reranking
voyageai>=0.3.0

# Environment and config
python-dotenv>=1.0.0

# Async support
anyio>=3.7.0

# JWT Authentication
python-jose>=3.3.0

# Logging
loguru>=0.7.0

# Retry logic
tenacity>=8.0.0

# Optional: numpy for embeddings
numpy>=1.24.0
tika==3.1.0
